refactor: converted quicksort to ts
